Refactor: Frame, Wrapper 컴포넌트 app/_component로 이동
